==> notebooks/03_Format_for_datascape.ipynb <==
# FORMAT DATA FOR DATASCAPE

The data must have been prepared beforehand with the notebook "01_Prepare_Infomedia_Data.ipnyb" of the series. It needs a file generated by that notebook.

The formatted data is intended to be used with the datascapes from this repository: https://github.com/jacomyma/add-datascapes

*Authors: Snorre Ralund, Mathieu Jacomy, Anders Munk.*

## How to use
1. Open 'INFOMEDIA NER PMI NETWORK.graphml' in **Gephi**
1. **Render its layout** in Gephi. [Check this tutorial](https://www.youtube.com/watch?v=0LqY8OfSsKE).
1. **Export it as a GraphML file** named for instance 'INFOMEDIA NER PMI NETWORK LAYOUT.graphml'. Put it in the same folder as this notebook.
1. Edit the settings below
1. Run all the cells from the menu: Cell > Run all.
1. It's over when the last cell outputs "Done."


settings = {}

# Source file (LAYOUT MUST BE DONE IN GEPHI)
settings['named_entities_network'] = 'INFOMEDIA NER PMI NETWORK LAYOUT.graphml'


**The rest of the script does not require your intervention, except for checking that no errors occur.**

Although you are welcome to change it! That is why it is commented.


# Installs
! pip install networkx

# Imports
import networkx as nx
import pandas as pd

# Extract the basemap file from the network
G = nx.read_graphml(settings['named_entities_network'])
entities = []
for nid, n in G.nodes(data = True):
    ent = {'Id':nid, 'label':nid, 'x':att['x'], 'y':att['y'], 'size':att['size'], 'showlabel':'true'}
    entities.append(ent)
ent_df = pd.DataFrame(entities)
ent_df

# Save basemap file
ent_df.to_csv('basemap.csv', index=False)
print("Done.")
